requirements.txt
pytest==7.4.2
selenium==4.12.0
faker==19.11.0
allure-pytest==2.13.2

